requirements.txt
# Core web framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
aiofiles==24.1.0

# Image processing
pillow==10.4.0
opencv-python==4.10.0.84
numpy==2.1.3

# Machine Learning & Computer Vision
mediapipe==0.10.13
torch==2.6.0
torchvision==0.21.0
transformers==4.45.0
open-clip-torch==2.27.0
onnxruntime==1.20.0

# ML utilities
scikit-learn==1.5.2
scipy==1.14.1

